% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_ssFC.R
\name{compute_ssFC}
\alias{compute_ssFC}
\title{Compute single sample logFCs}
\usage{
compute_ssFC(logCPM, metadata, factor, control)
}
\arguments{
\item{logCPM}{Matrix of normaslised logCPM}

\item{metadata}{Sample metadata data frame as described in the details section.}

\item{factor}{The factor defines how samples can be put into matching pairs.}

\item{control}{The treatment level that is the control.}
}
\value{
A matrix of single sample logFC
}
\description{
Compute single sample logFCs for each treated samples using normalised logCPM values
}
\details{
The number of rows must equal to the number of columns in the logCPM matrix. It also must have a column called "sample" storing sample names, and a column called "treatment" storing treatment of each sample.
}
