% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normaliseByPermutation.R
\name{.generate_permutedFC}
\alias{.generate_permutedFC}
\title{Permute sample labels to generate permuted logFCs}
\usage{
.generate_permutedFC(logCPM, numOfTreat, NB = 1000, weight, seed)
}
\description{
Permute sample labels to generate permuted logFCs
}
