% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_GSnetwork.R
\name{get_GSgenelist}
\alias{get_GSgenelist}
\title{Get gene-set to gene_id data frame}
\usage{
get_GSgenelist(gsTopology)
}
\arguments{
\item{gsTopology}{List of pathway topology matrices generated using function \code{weightedAdjMatrix}}
}
\value{
dataframe
}
\description{
Get gene-set to gene_id data frame
}
