% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gsAnnotation_df}
\alias{gsAnnotation_df}
\title{gsAnnotation_df: Categorization of KEGG pathways used for community annotation}
\format{
A \code{data.frame} with 549 rows and 2 columns containing categorization of 549 KEGG pathways
\describe{
\item{gs_name}{Gene-set name}
\item{vategory}{Category}
}
}
\source{
\url{https://www.genome.jp/kegg/}
}
\usage{
data(gsAnnotation_df)
}
\description{
gsAnnotation_df: Categorization of KEGG pathways used for community annotation
}
\keyword{datasets}
