% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_GSnetwork.R
\name{make_gsNetwork}
\alias{make_gsNetwork}
\title{Create gene-set network plot object}
\usage{
make_gsNetwork(
  normalisedScores,
  gsTopology,
  colorBy = c("robustZ", "pvalue"),
  foldGSname = TRUE,
  foldafter = 2
)
}
\arguments{
\item{normalisedScores}{A dataframe as described in the details section}

\item{gsTopology}{List of pathway topology matrices generated using function \code{weightedAdjMatrix}}

\item{colorBy}{Choose to color nodes either by "robustZ" or "pvalue". A column must exist in the normalisedScores for the chosen parameter}

\item{foldGSname}{logical(1). Should long gene-set names be folded into two lines}

\item{foldafter}{The number of words after which gene-set names should be folded. Defaulted to 2}
}
\description{
Create gene-set network plot object
}
