% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{metadata_example}
\alias{metadata_example}
\title{metadata_example: Sample metadata for malignant breast cancer tumours PDE from
5 ER+ breast cancer tumour (GSE80098)}
\format{
A \code{data.frame} with 15 rows and 4 columns
\describe{
\item{patient}{patient N2-3, P4-6 }
\item{treatment}{treatment: Vehicle, R5020 or E2+R5020}
\item{PR}{progesterone receptor status}
\item{sample}{sample name, corresponding to column names of the logCPM matrix}
}
}
\source{
\url{https://www.ncbi.nlm.nih.gov/pmc/articles/PMC4928895/}
}
\usage{
data(metadata_example)
}
\description{
metadata_example: Sample metadata for malignant breast cancer tumours PDE from
5 ER+ breast cancer tumour (GSE80098)
}
\keyword{datasets}
