% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normaliseByPermutation.R
\name{normaliseByPermutation_R}
\alias{normaliseByPermutation_R}
\title{Title}
\usage{
normaliseByPermutation_R(
  logCPM,
  metadata,
  factor,
  control,
  seed = sample.int(1e+06, 1),
  NB = 1000,
  scores,
  filePath,
  weight
)
}
\arguments{
\item{weight}{}
}
\value{

}
\description{
Title
}
