% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genePertScore.R
\name{pathway_pert}
\alias{pathway_pert}
\title{Compute Single-sample Pathway-level Perturbation Score}
\usage{
pathway_pert(weightedFC, genePertScore)
}
\arguments{
\item{weightedFC}{A matrix of weighted single sample logFCs derived from function \code{weight_ss_fc()}}

\item{genePertScore}{List of gene-wise raw perturbation score matrices generated using function \code{raw_gene_pert()}}
}
\value{
A data.frame with 3 columns: tA (single-sample pathway-level perturbation score), sample, and gs_name (gene-set name)
}
\description{
Propagate weighted single sample logFCs down the pathway topologies to compute single sample perturbation scores for each pathway
}
\details{
This function use the algorithm adopted from \code{SPIA} (see citation) to compute a single sample perturbation score per sample per
pathway. The rownames of the weighted single sample logFC matrix and the pathway topology matrices must use the same type of gene identifier (ie. entrez ID).
}
\examples{
#compute weighted single sample logFCs
data(metadata_example)
data(logCPM_example)
ls <- weight_ss_fc(logCPM_example, metadata = metadata_example,
factor = "patient", control = "Vehicle")
# extract all the KEGG pathways
gsTopology <- retrieve_topology(database = "kegg")
# compute raw gene-wise perturbation scores
genePertScore <- raw_gene_pert(ls$logFC, gsTopology)
# sum gene-wise perturbation scores to derive the pathway-level single-sample perturbation scores
pathwayPertScore <- pathway_pert(ls$logFC, genePertScore)
# rank genes by gene-wise perturbation scores within each sample to compare their contributions to pathway perturbation
geneRank <- rank_gene_pert(ls$logFC,genePertScore,  gsTopology)
}
\references{
Tarca AL, Draghici S, Khatri P, Hassan SS, Mittal P, Kim JS, Kim CJ, Kusanovic JP, Romero R. A novel signaling pathway impact analysis.
Bioinformatics. 2009 Jan 1;25(1):75-82.
}
