% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normaliseByPermutation.R
\name{permutedScore_PARALLEL}
\alias{permutedScore_PARALLEL}
\title{Title}
\usage{
permutedScore_PARALLEL(permutedFC, gsTopology, ncores, tol = 1e-07)
}
\arguments{
\item{ncores}{}
}
\value{

}
\description{
Title
}
