% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normaliseByPermutation.R
\name{permutedScore_Rcpp_para}
\alias{permutedScore_Rcpp_para}
\title{Title}
\usage{
permutedScore_Rcpp_para(permutedFC, gsTopology, ncores, tol = 1e-07)
}
\arguments{
\item{ncores}{}
}
\value{

}
\description{
Title
}
