% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pertubationScore.R
\name{perturbationScore}
\alias{perturbationScore}
\title{Single Sample Perturbation Score}
\usage{
perturbationScore(weightedFC, filePath)
}
\arguments{
\item{weightedFC}{A matrix of weighted single sample logFCs derived from function \code{weight_ssFC}}

\item{filePath}{The file path to pathway topology matrices generated using function \code{weightedAdjMatrix}}
}
\value{
A list where each element is a matrix corresponding to a pathway. Each column of an element corresponds to a sample.
}
\description{
Propagate weighted single sample logFCs down the pathway topologies to compute single sample perturbation scores for each pathway
}
\details{
This function use the algorithm adopted from \code{SPIA} (see citation) to compute a single sample perturbation score per sample per
pathway. The rownames of the weighted single sample logFC matrix and the pathway toplogy matrices must use the same type of gene identifier.
}
\references{
Tarca AL, Draghici S, Khatri P, Hassan SS, Mittal P, Kim JS, Kim CJ, Kusanovic JP, Romero R. A novel signaling pathway impact analysis.
Bioinformatics. 2009 Jan 1;25(1):75-82.
}
