% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_community.R
\name{plot_community}
\alias{plot_community}
\title{Visualize community structure in the significantly perturbed gene-set network}
\usage{
plot_community(
  normalisedScores,
  gsTopology,
  gsAnnotation = NULL,
  colorBy = c("robustZ", "pvalue", "community"),
  communityMethod = "cluster_louvain",
  foldGSname = TRUE,
  foldafter = 2,
  layout = "fr",
  markCommunity = "ellipse",
  markAlpha = 0.2,
  edgeAlpha = 0.8,
  up_col = "brown3",
  down_col = "steelblue3",
  scale_edgeWidth = c(0.5, 3),
  edgeLegend = FALSE,
  scale_nodeSize = c(3, 6),
  nodeShape = 16,
  color_lg = TRUE,
  color_lg_title = NULL,
  lb_size = 3,
  lb_color = "black",
  plotIsolated = FALSE,
  ...
)
}
\arguments{
\item{normalisedScores}{A \code{data.frame} derived from the \code{normalise_by_permu()} function}

\item{gsTopology}{List of pathway topology matrices generated using function \code{retrieve_topology()}}

\item{gsAnnotation}{A \code{data.frame} containing gene-sets categories used for pathway annotation. Must contain at least two columns:
\code{c("gs_name", "category")}, where \code{gs_name} denotes gene-sets names that are matched to names of pathway topology matrices, and
\code{category} records the categorization of the given pathway. If customized annotation is not provided, it's assumed that the pathways
investigated were from KEGG database and the inbuilt KEGG pathway annotation information will be used,}

\item{colorBy}{Choose to color nodes either by \emph{"community}, \emph{robustZ} or \emph{pvalue}. To color by \emph{robustZ} or \emph{pvalue}, a
column must exist in the normalisedScores for the chosen parameter}

\item{communityMethod}{A community Detection method supported by \code{igraph}. See details for all methods available.}

\item{foldGSname}{\code{logical}. Should long gene-set names be folded into two lines}

\item{foldafter}{The number of words after which gene-set names should be folded. Defaulted to 2}

\item{layout}{The layout algorithm to apply. Accept all layout supported by \code{igraph}. See details for all layout options available.}

\item{markCommunity}{\code{character} A \emph{geom_mark_} method supported by \code{ggforce} to annotate sets of nodes belonging to the same community.
Either \emph{NULL}, \emph{ellipse}, \emph{circle}, \emph{hull}, \emph{rect}}

\item{markAlpha}{Transparency of annotation areas. Default to 0.2}

\item{edgeAlpha}{Transparency of edges. Default to 0.8}

\item{up_col}{The color used to label activated gene-sets. Only applicable if \code{colorBy} is set to be "robustZ"}

\item{down_col}{The color used to label inhibited gene-sets. Only applicable if \code{colorBy} is set to be "robustZ"}

\item{scale_edgeWidth}{A numerical vector of length 2 to be provided to \code{ggraph::scale_edge_width_continuous()} for specifying
the minimum and maximum edge widths after transformation. Defaulted to c(0.5, 3)}

\item{edgeLegend}{logical` Should edge weight legend be shown}

\item{scale_nodeSize}{A numerical vector of length 2 to be provided to \code{ggplot2::scale_size()} for specifying
the minimum and maximum node sizes after transformation. Defaulted to c(3,6)}

\item{nodeShape}{The shape to use for nodes}

\item{color_lg}{\code{logical} Should color legend be shown}

\item{color_lg_title}{Title for the color legend}

\item{lb_size}{Size of node text labels}

\item{lb_color}{Color of node text labels}

\item{plotIsolated}{\code{logical}.Should nodes not connected to any other nodes be plotted.  Default to FALSE}

\item{...}{Used to pass various potting parameters to \verb{ggforce::geom_mark_*()}}
}
\value{
A ggplot2 object
}
\description{
Visualize community structure in the significantly perturbed gene-set network
}
\examples{
load(system.file("extdata", "gsTopology.rda", package = "sSNAPPY"))
load(system.file("extdata", "normalisedScores.rda", package = "sSNAPPY"))
#Subset the first 10 rows of the normalisedScores data.frame as an example
subset <- normalisedScores[1:15,]
# Color network plot nodes by the community they were assigned to and mark nodes belonging
# to the same community by ellipses
plot_community(subset, gsTopology, colorBy = "community",layout = "kk",
color_lg_title = "Community")

# Color network plot nodes by pathways' directions of changes and mark nodes belonging
# to the same community by ellipses
plot_community(subset, gsTopology, colorBy = "robustZ",layout = "kk",
color_lg_title = "Direction of pathway Perturbation")

# Plotting parameters accepted by `geom_mark_*` could be passed to the function
# to adjust the annotation area or the annotation label. See
# `?ggforce::geom_mark_*` for more details. For example, to change the linetype
# of the connector:
plot_community(subset, gsTopology, colorBy = "robustZ",layout = "kk",
color_lg_title = "Direction of pathway Perturbation", con.linetype = "dashed")

# To change the colour and fill of `geom_mark_*` annotation, use any
# `scale_fill_*` and/or `scale_color_*`
# functions supported by `ggplot2`. For example:
p <- plot_community(subset, gsTopology, colorBy = "robustZ",layout = "kk",
markCommunity = "rect",color_lg_title = "Direction of pathway Perturbation")
p + ggplot2::scale_color_ordinal() + ggplot2::scale_fill_ordinal()
}
