% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gs2gene.R
\name{plot_gs2gene}
\alias{plot_gs2gene}
\title{Plot pathways and genes contained in them as a network}
\usage{
plot_gs2gene(
  normalisedScores,
  gsTopology,
  geneFC = NULL,
  mapEntrezID = NULL,
  colorGS_By = c("robustZ", "pvalue"),
  foldGSname = TRUE,
  foldafter = 2,
  layout = "fr",
  edgeAlpha = 0.8,
  upGS_col = "brown3",
  downGS_col = "steelblue3",
  upGene_col = "pink",
  downGene_col = "lightblue",
  GeneNode_size = 3,
  GeneNode_shape = 17,
  GsNode_size = 2,
  GsNode_shape = 16,
  label_Gene = TRUE,
  GeneName_size = 3,
  GsName_size = 6,
  gene_lg_title = "Changes in Gene Expression",
  gs_lg_title = "Pathway Perturbation",
  arc_strength = 0.5
)
}
\arguments{
\item{normalisedScores}{A \code{data.frame} derived from the \code{normalise_by_permu()} function. Only gene-sets of interest should be included}

\item{gsTopology}{List of pathway topology matrices generated using function \code{retrieve_topology()}}

\item{geneFC}{An optional named vector of pathways' fold changes}

\item{mapEntrezID}{Optional. A \code{data.frame} matching genes' entrez IDs to other identifier. Must contain 2 columns: \code{"entrezid"} and \code{"mapTo"}}

\item{colorGS_By}{Choose to color nodes by \emph{robustZ} or \emph{pvalue}. A column must exist in the normalisedScores \code{data.frame}
for the chosen parameter}

\item{foldGSname}{\code{logical}. Should long gene-set names be folded into two lines}

\item{foldafter}{The number of words after which gene-set names should be folded. Defaulted to 2}

\item{layout}{The layout algorithm to apply. Accept all layout supported by \code{igraph}.}

\item{edgeAlpha}{Transparency of edges. Default to 0.8}

\item{upGS_col}{Color for activated gene-sets. Only applicable if \code{colorGS_By} is set to be "robustZ"}

\item{downGS_col}{Color for inhibited gene-sets. Only applicable if \code{colorGS_By} is set to be "robustZ"}

\item{upGene_col}{Color for up-regulated genes. Only applicable if \code{geneFC} is not NULL}

\item{downGene_col}{Color for down-regulated genes. Only applicable if \code{geneFC} is not NULL}

\item{GeneNode_size}{Size for gene nodes}

\item{GeneNode_shape}{Shape for gene nodes}

\item{GsNode_size}{Size for gene-set nodes}

\item{GsNode_shape}{Shape for gene nodes}

\item{label_Gene}{\code{logical} Should gene name be plotted}

\item{GeneName_size}{Size of gene name label}

\item{GsName_size}{Size of gene-set name label}

\item{gene_lg_title}{`character. Legend for gene nodes color}

\item{gs_lg_title}{\code{character} Legend for gene-set nodes color}

\item{arc_strength}{The bend of edges. 1 approximates a halfcircle while 0 will give a straight line.}
}
\value{
A ggplot2 object
}
\description{
Plot pathways and genes contained in them as a network
}
\details{
Taking the perturbation scores of a list of gene-sets derived from \code{normalise_by_permu()} as input, this function matches gene-set to
their associated genes by utilizing information from pathway topology matrices.

It's optional to provide genes' logFCs as a named vector, where the names must be genes' entrez IDs in the format of "ENTREZID:XXXX".
This is because pathway topology matrices retrieved through \code{retrieve_topology()} always use entrez ID as identifiers.

However, it might not be very informative to label genes with their entrez ID. So users can also choose to proivde a \code{mapEntrezID}
\code{data.frame} to match entrez IDs to their chosen identifiers. The \code{data.frame} should contain two columns: \code{"entrezid"} and \code{"mapTo"}.

If \code{geneFC} is provided, gene nodes will be colored by genes' directions of changes. Otherwise, all gene nodes will be black.

Since some gene-sets could can contain hundreds of genes, it is not recommended to plot all of those genes. If \code{mapEntrezID} \code{data.frame} is provided,
only genes included in that \code{data.frame} will be used in the plot.

It is strongly recommended to filter for genes with the highest magnitude of changes. If all pathway genes
have to be plotted, consider setting \code{label_Gene} to FALSE to turn off plotting all gene names.
}
\examples{
load(system.file("extdata", "gsTopology.rda", package = "sSNAPPY"))
load(system.file("extdata", "normalisedScores.rda", package = "sSNAPPY"))
#Subset pathways significantly perturbed in sample R5020_N2_48
subset <- dplyr::filter(normalisedScores, adjPvalue < 0.05, sample == "R5020_N2_48")

# Color gene-sets nodes by robust z-scores.
plot_gs2gene(subset, gsTopology, colorGS_By = "robustZ", label_Gene = FALSE,
GeneNode_size = 1)
# When genes' fold-changes are not provided, gene nodes are colored in black.

# To color genes by their directions of changes, firstly compute genes' single-sample logFCs
data(logCPM_example)
data(metadata_example)
ls <- weight_ss_fc(logCPM_example, metadata = metadata_example,
 factor = "patient", control = "Vehicle")
# Provide fold-changes of sample R5020_N2_48 as a named vector
plot_gs2gene(subset, gsTopology, geneFC = ls$logFC[,"R5020_N2_48"], colorGS_By = "robustZ",
label_Gene = FALSE)

# There are still a large number of genes, making the plot cumbersome. There only fold-changes of
# genes with top 500 absolute fold-changes are provide so only pathway genes in that list of 500
# genes were plotted.
FC <- sort(abs(ls$logFC[,"R5020_N2_48"]), decreasing = TRUE)[1:500]
plot_gs2gene(subset, gsTopology, geneFC = FC, colorGS_By = "robustZ")

# To make the gene labels more informative, map genes' entrez id to chosen identifiers.
load(system.file("extdata", "entrez2name.rda", package = "sSNAPPY"))
plot_gs2gene(subset, gsTopology, geneFC = FC, mapEntrezID = entrez2name, colorGS_By = "robustZ")
}
