% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_GSnetwork.R
\name{plot_gsNetwork}
\alias{plot_gsNetwork}
\title{Plot gene-set network}
\usage{
plot_gsNetwork(
  normalisedScores,
  gsTopology,
  colorBy = c("robustZ", "pvalue"),
  foldGSname = TRUE,
  foldafter = 2,
  layout = "fr",
  edgeAlpha = 0.8,
  up_col = "brown3",
  down_col = "steelblue3",
  scale_edgeWidth = 10,
  scale_nodeSize = 15,
  nodeShape = 16,
  color_lg = TRUE,
  color_lg_title = NULL,
  lb_size = 3,
  lb_color = "black"
)
}
\arguments{
\item{normalisedScores}{A dataframe as described in the details section}

\item{gsTopology}{List of pathway topology matrices generated using function \code{weightedAdjMatrix}}

\item{colorBy}{Choose to color nodes either by "robustZ" or "pvalue". A column must exist in the normalisedScores for the chosen parameter}

\item{foldGSname}{logical(1). Should long gene-set names be folded into two lines}

\item{foldafter}{The number of words after which gene-set names should be folded. Defaulted to 2}

\item{layout}{The layout algorithm to apply}

\item{edgeAlpha}{Transparency of edges}

\item{up_col}{The color to label activated gene-sets. Only applicable if \code{colorBy} is set to be "robustZ"}

\item{down_col}{The color to label inhibited gene-sets. Only applicable if \code{colorBy} is set to be "robustZ"}

\item{scale_edgeWidth}{Parameter for scaling edge width. Defaulted to 10. Higher numbers reduce all edge width}

\item{scale_nodeSize}{Parameter for scaling node size. Defaulted to 15. Higher numbers decreases all node sizes}

\item{nodeShape}{The shape to use for nodes}

\item{color_lg}{logical(1). Should color legend be shown}

\item{color_lg_title}{Title for the color legend}

\item{lb_size}{Size of node text labels}

\item{lb_color}{Color of node text labels}
}
\value{
A ggplot2 object
}
\description{
Plot gene-set network
}
\examples{
load(system.file("extdata", "gsTopology.rda", package = "sSNAPPY"))
load(system.file("extdata", "normalisedScores.rda", package = "sSNAPPY"))
# Extract pathways significantly perturbed in a sample
normScores_sub <- dplyr::filter(normalisedScores, sample == "E2+R5020_N2_48", adjPvalue < 0.05)

# Color network plot nodes by robust z-score
plot_gsNetwork(normScores_sub, gsTopology,
colorBy = "robustZ", layout = "dh",
color_lg_title = "Robust Z-score")

# Color network plot nodes by p-values
plot_gsNetwork(normScores_sub, gsTopology, layout = "dh",
colorBy = "pvalue", color_lg_title = "P-value")
}
