% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genePertScore.R
\name{rank_gene_pert}
\alias{rank_gene_pert}
\title{Rank genes by perturbation contributions within each sample}
\usage{
rank_gene_pert(weightedFC, genePertScore, gsTopology)
}
\arguments{
\item{weightedFC}{A matrix of weighted single sample logFCs derived from function \code{weight_ss_fc()}}

\item{genePertScore}{List of gene-wise raw perturbation score matrices generated using function \code{raw_gene_pert()}}

\item{gsTopology}{List of pathway topology matrices generated using function \code{retrieve_topology()}}
}
\value{
A list where each element is a matrix corresponding to a pathway. Each column of an element corresponds to a sample, and each row corresponds to a pathway gene.
}
\description{
Rank genes by single-sample gene-wise raw perturbation scores to compare genes' contributions to pathway perturbation.
}
\details{
Ranking is performed within each sample each pathway. If in a given pathway, both positive and negative gene-wise perturbation scores exist, positive
and negative scores are ranked separately, where the larger a positive rank, the more the gene contributed to the pathway's activation, and the smaller a negative
rank, the more the gene contributed to the pathways' inhibition. When there's a tie in two gene's perturbation score within a sample, the mean of the indices is used.
}
\examples{
#compute weighted single sample logFCs
data(metadata_example)
data(logCPM_example)
ls <- weight_ss_fc(logCPM_example, metadata = metadata_example,
factor = "patient", control = "Vehicle")
# extract all the KEGG pathways
gsTopology <- retrieve_topology(database = "kegg")
# compute raw gene-wise perturbation scores
genePertScore <- raw_gene_pert(ls$logFC, gsTopology)
# rank genes by gene-wise perturbation scores within each sample to compare their contributions to pathway perturbation
geneRank <- rank_gene_pert(ls$logFC,genePertScore,  gsTopology)
}
\references{
Tarca AL, Draghici S, Khatri P, Hassan SS, Mittal P, Kim JS, Kim CJ, Kusanovic JP, Romero R. A novel signaling pathway impact analysis.
Bioinformatics. 2009 Jan 1;25(1):75-82.
}
