% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SSPT.R
\docType{package}
\name{sSNAPPY}
\alias{sSNAPPY}
\title{sSNAPPY: A package for testing directional single sample pathway perturbation}
\description{
A package for testing directional single sample pathway perturbation
}
