% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_GSnetwork.R
\name{str_replace_nth}
\alias{str_replace_nth}
\title{Replace the nth occurrence of string with another string}
\usage{
str_replace_nth(x, pattern, replacement, n)
}
\arguments{
\item{x}{string to search for pattern}

\item{pattern}{pattern to search for}

\item{replacement}{replacement string}

\item{n}{the nth orrcurrences of pattern that will be replaced}
}
\value{
string
}
\description{
Replace the nth occurrence of string with another string
}
