% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pertubationScore.R
\name{weightedAdjMatrix}
\alias{weightedAdjMatrix}
\title{Create weighted adjacent matrix}
\usage{
weightedAdjMatrix(list, beta = NULL, filename)
}
\arguments{
\item{list}{Pathway topology information retrieved using \code{graphite}. See example}

\item{beta}{A named numeric vector of weights to be assigned to each type of gene/protein relation type. See details for more information.}

\item{filename}{A file directory specifying where the output should be stored.}
}
\description{
Convert pathway topology matrices to normalized weighted directed adjacency matrices describing the gene signaling network.
}
\details{
This function takes the pathway topology information retrieved using \code{graphite} and convert them to normalized weighted directed adjacency
matrices describing the gene signaling network. See cited document for more details.

The beta parameter speicifies weights to be assigned to each type of gene-gene interaction. It should be a named numeric vector of length 23,
whose names must be: c("activation","compound","binding/association","expression","inhibition","activation_phosphorylation","phosphorylation",
"indirect","inhibition_phosphorylation","dephosphorylation_inhibition","dissociation","dephosphorylation","activation_dephosphorylation",
"state","activation_indirect","inhibition_ubiquination","ubiquination","expression_indirect","indirect_inhibition",
"repression","binding/association_phosphorylation","dissociation_phosphorylation","indirect_phosphorylation"). If unspecified, beta will be by default
chosen as: c(1,0,0,1,-1,1,0,0,-1,-1,0,0,1,0,1,-1,0,1,-1,-1,0,0,0).
}
\references{
Tarca AL, Draghici S, Khatri P, Hassan SS, Mittal P, Kim JS, Kim CJ, Kusanovic JP, Romero R. A novel signaling pathway impact analysis.
Bioinformatics. 2009 Jan 1;25(1):75-82.
}
