% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normaliseByPermutation.R
\name{.generate_permutedFC}
\alias{.generate_permutedFC}
\title{Title}
\usage{
.generate_permutedFC(logCPM, metadata, factor, control, weight, NB)
}
\arguments{
\item{BPPARAM}{}
}
\value{

}
\description{
Title
}
