% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normaliseByPermutation.R
\name{.generate_permutedFC}
\alias{.generate_permutedFC}
\title{#' Title
#'
#' @param logCPM
#' @param metadata
#' @param factor
#' @param control
#' @param BPPARAM
#' @param seed
#'
#' @return
#'
#' @examples}
\usage{
.generate_permutedFC(
  logCPM,
  metadata,
  factor,
  control,
  weight,
  NB,
  seed,
  ncores = 6
)
}
\description{
#' Title
#'
#' @param logCPM
#' @param metadata
#' @param factor
#' @param control
#' @param BPPARAM
#' @param seed
#'
#' @return
#'
#' @examples
}
