% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{metadata_example}
\alias{metadata_example}
\title{metadata_example: Sample metadata for malignant breast cancer tumors PDE from 8 ER+PR+ breast cancer patients (GSE80098)}
\format{
A data frame with 16 rows and 6 columns
\describe{
\item{patient}{patient p1-p8}
\item{treatment}{treatment: Vehicle or E2}
\item{time}{amount of time the sample was treated for, in hours}
\item{sample}{sample name, corresponding to column names of the logCPM matrix}
}
}
\source{
\url{https://www.ncbi.nlm.nih.gov/pmc/articles/PMC4928895/}
}
\usage{
metadata_example
}
\description{
metadata_example: Sample metadata for malignant breast cancer tumors PDE from 8 ER+PR+ breast cancer patients (GSE80098)
}
\keyword{datasets}
