% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normaliseByPermutation.R
\name{normaliseByPermutation}
\alias{normaliseByPermutation}
\title{Normalise test perturbation scores by permutation}
\usage{
normaliseByPermutation(permutedScore, testScore, pAdj_method = "fdr")
}
\arguments{
\item{permutedScore}{A list. Output of \code{generate_PermutedScore}}

\item{testScore}{A matrix. Output of \code{weight_ssFC}}

\item{pAdj_method}{Method for adjusting p-values for multiple comparisons. See \code{?p.adjust} for methods available. Default to FDR.}
}
\value{
A dataframe.
}
\description{
Normalise test perturbation scores by permutation
}
\details{
Normalise the test perturbation scores generated by \code{weight_ssFC} through the permuted perturbation scores derived from
the \code{generate_PermutedScore} function. The mean absolute deviation(MAD) and median of perturbation scores for each pathway are firstly
derived from the permuted perturbation scores. The test perturbation scores are then converted to robust z-scores using MADs and medians
calculated.
}
\examples{
\dontrun{
load(system.file("extdata", "gsTopology.rda", package = "sSNAPPY"))
ssPertScore <- perturbationScore(ls$logFC, gsTopology)
permutedScore <- generate_PermutedScore(logCPM_example, numOfTreat = 2,
 NB = 1000, gsTopology = gsTopology, weight = ls$weight)
normalisedScores <- normaliseByPermutation(permutedScore, ssPertScore)
 }
 # Load the example output
 load(system.file("extdata", "normalisedScores.rda", package = "sSNAPPY"))
 normalisedScores
}
