% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pertubationScore.R
\name{perturbationScore}
\alias{perturbationScore}
\title{Single Sample Perturbation Score}
\usage{
perturbationScore(weightedFC, filePath)
}
\arguments{
\item{weightedFC}{A matrix of weighted single sample logFCs derived from function \code{weight_ssFC}}

\item{filePath}{The file path to pathway topology matrices generated using function \code{weightedAdjMatrix}}
}
\value{
A list where each element is a matrix corresponding to a pathway. Each column of an element corresponds to a sample.
}
\description{
Propagate weighted single sample logFCs down the pathway topologies to compute single sample perturbation scores for each pathway
}
\details{
This function use the algorithm adopted from \code{SPIA} (see citation) to compute a single sample perturbation score per sample per
pathway. The rownames of the weighted single sample logFC matrix and the pathway toplogy matrices must use the same type of gene identifier (ie. entrez ID).
}
\examples{
require(AnnotationHub)
# convert rownamews of logCPM from gene ids to gene entrez IDs through `AnnotationHub`
ah <- AnnotationHub()
ah <- subset(ah,genome == "GRCh38" & title == "Ensembl 101 EnsDb for Homo sapiens")
ensDb <- ah[[1]]
rownames(logCPM_example) <- AnnotationHub::mapIds(ensDb, rownames(logCPM_example), "ENTREZID", keytype = "GENEID")

# Remove genes that couldn't be matched to entrez IDs
logCPM_example <- logCPM_example[!is.na(rownames(logCPM_example)),]

#compute weighted single sample logFCs
ls <- weight_ssFC(logCPM_example, metadata = metadata_example, factor = "patient", control = "Vehicle")

# explore all species and databases supported by graphite
graphite::pathwayDatabases()
weightedAdjMatrix(species = "hsapiens", database = "kegg", outputDir = "BminsI.rds")

ssPertScore <- perturbationScore(ls$logFC, filePath = "BminsI.rds")
}
\references{
Tarca AL, Draghici S, Khatri P, Hassan SS, Mittal P, Kim JS, Kim CJ, Kusanovic JP, Romero R. A novel signaling pathway impact analysis.
Bioinformatics. 2009 Jan 1;25(1):75-82.
}
