% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gene_contribution.R
\name{plot_gene_contribution}
\alias{plot_gene_contribution}
\title{Plot genes' contribution to pathway perturbation as heatmap}
\usage{
plot_gene_contribution(
  geneRankList,
  gsToPlot,
  rownames,
  metadata = NULL,
  annotation_attribute = c("pathwayPertScore"),
  pathwayPertScore = NULL,
  colOrder
)
}
\arguments{
\item{geneRankList}{List of gene-wise rank matrices generated using function \code{raw_gene_pert()}}

\item{gsToPlot}{Name of pathway to be plotted}

\item{metadata}{Sample metadata data frame containing information for heatmap annotation}

\item{annotation_attribute}{Characteristic vector specifying attributes to draw annotation for. Default to "pathwayPertScore" (
ie. pathway-level perturbation scores)}

\item{pathwayPertScore}{A dataframe. Output of function \code{pathway_pert()}}

\item{colOrder}{}
}
\description{
Plot genes' contribution to pathway perturbation as heatmap
}
\examples{
#compute weighted single sample logFCs
data(metadata_example)
data(logCPM_example)
ls <- weight_ss_fc(logCPM_example, metadata = metadata_example,
factor = "patient", control = "Vehicle")
# extract all the KEGG pathways
gsTopology <- retrieve_topology(database = "kegg")
# compute raw gene-wise perturbation scores
genePertScore <- raw_gene_pert(ls$logFC, gsTopology)
# rank genes by gene-wise perturbation scores within each sample to compare their contributions to pathway perturbation
geneRank <- rank_gene_pert(ls$logFC,genePertScore,  gsTopology)
# sum gene-wise perturbation scores to derive the pathway-level single-sample perturbation scores
pathwayPertScore <- pathway_pert(ls$logFC, genePertScore)
plot_gene_contribution(geneRankList = geneRank, gsToPlot = "Estrogen signaling pathway")
}
