% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_GSnetwork.R
\name{plot_gs_network}
\alias{plot_gs_network}
\title{Plot gene-set network}
\usage{
plot_gs_network(
  normalisedScores,
  gsTopology,
  colorBy = c("robustZ", "pvalue"),
  foldGSname = TRUE,
  foldafter = 2,
  layout = "fr",
  edgeAlpha = 0.8,
  up_col = "brown3",
  down_col = "steelblue3",
  scale_edgeWidth = c(0.5, 3),
  edgeLegend = FALSE,
  scale_nodeSize = c(3, 6),
  nodeShape = 16,
  color_lg = TRUE,
  color_lg_title = NULL,
  lb_size = 3,
  lb_color = "black",
  plotIsolated = FALSE
)
}
\arguments{
\item{normalisedScores}{A dataframe as described in the details section}

\item{gsTopology}{List of pathway topology matrices generated using function \code{retrieve_topology}}

\item{colorBy}{Choose to color nodes either by "robustZ" or "pvalue". A column must exist in the normalisedScores for the chosen parameter}

\item{foldGSname}{logical(1). Should long gene-set names be folded into two lines}

\item{foldafter}{The number of words after which gene-set names should be folded. Defaulted to 2}

\item{layout}{The layout algorithm to apply}

\item{edgeAlpha}{Transparency of edges}

\item{up_col}{The color to label activated gene-sets. Only applicable if \code{colorBy} is set to be "robustZ"}

\item{down_col}{The color to label inhibited gene-sets. Only applicable if \code{colorBy} is set to be "robustZ"}

\item{scale_edgeWidth}{A numerical vector of length 2 to be provided to \code{ggraph::scale_edge_width_continuous()} for specifying
the minimum and maximum edge widths after transformation. Defaulted to c(0.5, 3)}

\item{scale_nodeSize}{A numerical vector of length 2 to be provided to \code{ggplot2::scale_size()} for specifying
the minimum and maximum node sizes after transformation. Defaulted to c(3,6)}

\item{nodeShape}{The shape to use for nodes}

\item{color_lg}{logical(1). Should color legend be shown}

\item{color_lg_title}{Title for the color legend}

\item{lb_size}{Size of node text labels}

\item{lb_color}{Color of node text labels}

\item{plotIsolated}{logical(1). If nodes not connected to any other node should be plotted. Default to FALSE}
}
\value{
A ggplot2 object
}
\description{
Plot gene-set network
}
\examples{
load(system.file("extdata", "gsTopology.rda", package = "sSNAPPY"))
load(system.file("extdata", "normalisedScores.rda", package = "sSNAPPY"))
#Subset pathways significantly perturbed in sample R5020_N2_48
subset <- dplyr::filter(normalisedScores, adjPvalue < 0.05, sample == "R5020_N2_48")
# Color network plot nodes by robust z-score
plot_gs_network(subset, gsTopology,
colorBy = "robustZ", layout = "dh",
color_lg_title = "Robust Z-score")

# Color network plot nodes by p-values
plot_gs_network(subset, gsTopology, layout = "dh",
colorBy = "pvalue", color_lg_title = "P-value")
}
