% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_ssFC.R
\name{weight_ssFC}
\alias{weight_ssFC}
\title{Compute weighted single sample LogFCs from normalised logCPM}
\usage{
weight_ssFC(logCPM, metadata, factor, control)
}
\arguments{
\item{logCPM}{Matrix of normaslised logCPM where rows are genes and columns are samples. Rownames need to be gene entrez ID.}

\item{metadata}{Sample metadata data frame as described in the details section.}

\item{factor}{The factor defines how samples can be put into matching pairs.}

\item{control}{The treatment level that is the control.}
}
\value{
A list
}
\description{
Compute weighted single sample logFCs for each treated samples using normalised logCPM values. Fit a lowess
curve on variance of single sample logFCs ~ mean of logCPM, and use this to predict a gene-wise weight. The weighted
single sample logFCs are ready for computing perturbation scores.
}
\details{
This function computes weighted single sample logFCs from normalised logCPM values, used for computing single sample perturbation scores.
Since genes with smaller logCPM turn to have a larger variance among single sample logFCs. A lowess curve is fitted to estimate the
relationship between variance of single sample logFCs and mean of logCPM, and use this relationship to estimate the variance of each
mean logCPM value. Weights, which are inverse variance, are then multiplied to single sample logFCs to downweight genes with low counts.

It is assumed that the genes with extremely low counts have been removed and the count matrix has been normalised prior to logCPM
matrix was derived. Rownames of the matrix must be genes' entrez ID. To convert ensemble ID to entrz ID, see example.

The number of rows in the sample metadata must equal to the number of columns in the logCPM matrix.
Metadata also must have a column called "sample" storing sample names, and a column called "treatment" storing
treatment of each sample.The control treatment level specified by \code{control} parameter must exist in the treatment column.

This analysis was designed for experimental designs that include matched pairs of samples, such as when the tissues collected from the
same patient were treated with different treatments. Parameter \code{factor} tells the function how samples can be put into pairs. It must also
be a column of the metadata.
}
