% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathwayTopology.R
\name{weightedAdjMatrix}
\alias{weightedAdjMatrix}
\title{Create pathway topology weighted adjacent matrix}
\usage{
weightedAdjMatrix(
  species,
  database,
  pathwayName = NULL,
  beta = NULL,
  outputDir
)
}
\arguments{
\item{species}{See example for supported species.}

\item{database}{See example for supported databases.}

\item{pathwayName}{Optional. Subset of pathway names as a vector.}

\item{beta}{Optional. A named numeric vector of weights to be assigned to each type of gene/protein relation type.
See details for more information.}

\item{outputDir}{A file directory specifying where the weighted adjacent matrix should be stored.}
}
\description{
Convert pathway topology matrices to normalized weighted directed adjacency matrices describing the gene signaling network.
}
\details{
This function takes the pathway topology information retrieved using \code{graphite} and convert them to normalized weighted directed adjacency
matrices describing the gene signaling network, which can be directed used to compute pathway perturbation score through \emph{SPIA} algorithm.
See cited document for more details.

The beta parameter specifies weights to be assigned to each type of gene-gene interaction. It should be a named numeric vector of length 23,
whose names must be: c("activation","compound","binding/association","expression","inhibition","activation_phosphorylation","phosphorylation",
"indirect","inhibition_phosphorylation","dephosphorylation_inhibition","dissociation","dephosphorylation","activation_dephosphorylation",
"state","activation_indirect","inhibition_ubiquination","ubiquination","expression_indirect","indirect_inhibition",
"repression","binding/association_phosphorylation","dissociation_phosphorylation","indirect_phosphorylation"). If unspecified, beta will be by default
chosen as: c(1,0,0,1,-1,1,0,0,-1,-1,0,0,1,0,1,-1,0,1,-1,-1,0,0,0).

The converted weighted adjacent matrices will be stored in list and write into the file directory specified through the \code{outputDir} parameter.
}
\examples{
# explore all species and databases supported by graphite
graphite::pathwayDatabases()
weightedAdjMatrix(species = "hsapiens", database = "kegg", outputDir = "BminsI.rds")

# if only interested in selected pathways, specify the pathway names in the `pathwayName` parameter
weightedAdjMatrix(species = "hsapiens", database = "kegg",
pathwayName = c("Glycolysis / Gluconeogenesis","Citrate cycle (TCA cycle)","Pentose phosphate pathway"),
outputDir = "BminsI.rds")
}
\references{
Tarca AL, Draghici S, Khatri P, Hassan SS, Mittal P, Kim JS, Kim CJ, Kusanovic JP, Romero R. A novel signaling pathway impact analysis.
Bioinformatics. 2009 Jan 1;25(1):75-82.
Sales, G., Calura, E., Cavalieri, D. et al. graphite - a Bioconductor package to convert pathway topology to gene network.
BMC Bioinformatics 13, 20 (2012).
}
